Split test_helper in several files
